Create README.md
The Data Science and Machine Learning for Asset Management Specialization has been designed to deliver a broad and comprehensive introduction to modern methods in Investment Management, with a particular emphasis on the use of data science and machine learning techniques to improve investment decisions.By the end of this specialization, you will have acquired the tools required for making sound investment decisions, with an emphasis not only on the foundational theory and underlying concepts, but also on practical applications and implementation. Instead of merely explaining the science, we help you build on that foundation in a practical manner, with an emphasis on the hands-on implementation of those ideas in the Python programming language through a series of dedicated lab sessions.